Upgrade offline routing sample
