+++ src/GeocodeAndRoutingOnMouseMove/GeocodeAndRoutingOnMouseMove.sln
@@ -1,24 +1,22 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.4.33110.190
+# Visual Studio Version 18
+VisualStudioVersion = 18.4.11626.88
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LocalNetworkSample.Desktop", "LocalNetworkSample.Desktop\LocalNetworkSample.Desktop.csproj", "{476E5E0B-479A-4ED3-9149-9C2F330F8442}"
 EndProject
 Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "LocalNetworkSample.Shared", "LocalNetworkSample.Shared\LocalNetworkSample.Shared.shproj", "{7CFECAEC-45CE-46FA-9F9A-7A0108DFA492}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LocalNetworkSample.UWP", "LocalNetworkSample.UWP\LocalNetworkSample.UWP.csproj", "{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LocalNetworkSample.WinUI", "LocalNetworkSample.WinUI\LocalNetworkSample.WinUI.csproj", "{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Debug|ARM = Debug|ARM
 		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
 		Release|ARM = Release|ARM
 		Release|x64 = Release|x64
-		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{476E5E0B-479A-4ED3-9149-9C2F330F8442}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
@@ -27,16 +25,12 @@ Global
 		{476E5E0B-479A-4ED3-9149-9C2F330F8442}.Debug|ARM.Build.0 = Debug|Any CPU
 		{476E5E0B-479A-4ED3-9149-9C2F330F8442}.Debug|x64.ActiveCfg = Debug|Any CPU
 		{476E5E0B-479A-4ED3-9149-9C2F330F8442}.Debug|x64.Build.0 = Debug|Any CPU
-		{476E5E0B-479A-4ED3-9149-9C2F330F8442}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{476E5E0B-479A-4ED3-9149-9C2F330F8442}.Debug|x86.Build.0 = Debug|Any CPU
 		{476E5E0B-479A-4ED3-9149-9C2F330F8442}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{476E5E0B-479A-4ED3-9149-9C2F330F8442}.Release|Any CPU.Build.0 = Release|Any CPU
 		{476E5E0B-479A-4ED3-9149-9C2F330F8442}.Release|ARM.ActiveCfg = Release|Any CPU
 		{476E5E0B-479A-4ED3-9149-9C2F330F8442}.Release|ARM.Build.0 = Release|Any CPU
 		{476E5E0B-479A-4ED3-9149-9C2F330F8442}.Release|x64.ActiveCfg = Release|Any CPU
 		{476E5E0B-479A-4ED3-9149-9C2F330F8442}.Release|x64.Build.0 = Release|Any CPU
-		{476E5E0B-479A-4ED3-9149-9C2F330F8442}.Release|x86.ActiveCfg = Release|Any CPU
-		{476E5E0B-479A-4ED3-9149-9C2F330F8442}.Release|x86.Build.0 = Release|Any CPU
 		{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}.Debug|Any CPU.ActiveCfg = Debug|x86
 		{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}.Debug|Any CPU.Build.0 = Debug|x86
 		{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}.Debug|Any CPU.Deploy.0 = Debug|x86
@@ -46,9 +40,6 @@ Global
 		{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}.Debug|x64.ActiveCfg = Debug|x64
 		{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}.Debug|x64.Build.0 = Debug|x64
 		{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}.Debug|x64.Deploy.0 = Debug|x64
-		{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}.Debug|x86.ActiveCfg = Debug|x86
-		{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}.Debug|x86.Build.0 = Debug|x86
-		{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}.Debug|x86.Deploy.0 = Debug|x86
 		{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}.Release|Any CPU.ActiveCfg = Release|x86
 		{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}.Release|Any CPU.Build.0 = Release|x86
 		{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}.Release|Any CPU.Deploy.0 = Release|x86
@@ -58,9 +49,6 @@ Global
 		{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}.Release|x64.ActiveCfg = Release|x64
 		{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}.Release|x64.Build.0 = Release|x64
 		{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}.Release|x64.Deploy.0 = Release|x64
-		{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}.Release|x86.ActiveCfg = Release|x86
-		{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}.Release|x86.Build.0 = Release|x86
-		{CA543A0D-7567-4A93-A9C5-B1BBB1A0B995}.Release|x86.Deploy.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -71,6 +59,6 @@ Global
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
 		LocalNetworkSample.Shared\LocalNetworkSample.Shared.projitems*{476e5e0b-479a-4ed3-9149-9c2f330f8442}*SharedItemsImports = 5
 		LocalNetworkSample.Shared\LocalNetworkSample.Shared.projitems*{7cfecaec-45ce-46fa-9f9a-7a0108dfa492}*SharedItemsImports = 13
-		LocalNetworkSample.Shared\LocalNetworkSample.Shared.projitems*{ca543a0d-7567-4a93-a9c5-b1bbb1a0b995}*SharedItemsImports = 4
+		LocalNetworkSample.Shared\LocalNetworkSample.Shared.projitems*{ca543a0d-7567-4a93-a9c5-b1bbb1a0b995}*SharedItemsImports = 5
 	EndGlobalSection
 EndGlobal
