+++ src/GeocodeAndRoutingOnMouseMove/LocalNetworkSample.Shared/Common/BooleanToVisibilityConverter.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading.Tasks;
-#if NETFX_CORE
-using Windows.UI.Xaml;
-using Windows.UI.Xaml.Data;
+#if WINUI
+using Microsoft.UI.Xaml;
+using Microsoft.UI.Xaml.Data;
 #else
 using System.Windows.Data;
 using System.Windows;
@@ -16,7 +16,7 @@ namespace LocalNetworkSample.Common
 {
     public sealed class BooleanToVisibilityConverter : IValueConverter
     {
-#if NETFX_CORE
+#if WINUI
         public object Convert(object value, Type targetType, object parameter, string language)
 #else
        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
@@ -27,7 +27,7 @@ namespace LocalNetworkSample.Common
             else
                 return (value is bool && (bool)value) ? Visibility.Visible : Visibility.Collapsed;
         }
-#if NETFX_CORE
+#if WINUI
         public object ConvertBack(object value, Type targetType, object parameter, string language)
 #else
 		public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
