--- src/GeocodeAndRoutingOnMouseMove/LocalNetworkSample.UWP/Package.appxmanifest
@@ -1,34 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Package xmlns="http://schemas.microsoft.com/appx/manifest/foundation/windows10" xmlns:mp="http://schemas.microsoft.com/appx/2014/phone/manifest" xmlns:uap="http://schemas.microsoft.com/appx/manifest/uap/windows10" IgnorableNamespaces="uap mp">
-  <Identity Name="7d33c9ee-f2eb-48ba-8b78-406a20cc4ff9" Publisher="CN=Greg" Version="1.0.0.0" />
-  <mp:PhoneIdentity PhoneProductId="7d33c9ee-f2eb-48ba-8b78-406a20cc4ff9" PhonePublisherId="00000000-0000-0000-0000-000000000000" />
-  <Properties>
-    <DisplayName>LocalNetworkSample</DisplayName>
-    <PublisherDisplayName>Greg</PublisherDisplayName>
-    <Logo>Assets\StoreLogo.png</Logo>
-  </Properties>
-  <Dependencies>
-    <TargetDeviceFamily Name="Windows.Universal" MinVersion="10.0.0.0" MaxVersionTested="10.0.0.0" />
-  </Dependencies>
-  <Resources>
-    <Resource Language="x-generate" />
-  </Resources>
-  <Applications>
-    <Application Id="App" Executable="$targetnametoken$.exe" EntryPoint="LocalNetworkSample.App">
-      <uap:VisualElements DisplayName="Network Sample" Square150x150Logo="Assets\Logo.png" Square44x44Logo="Assets\SmallLogo.png" Description="LocalNetworkSample" BackgroundColor="transparent">
-        <uap:DefaultTile ShortName="Network Sample">
-          <uap:ShowNameOnTiles>
-            <uap:ShowOn Tile="square150x150Logo" />
-          </uap:ShowNameOnTiles>
-        </uap:DefaultTile>
-        <uap:SplashScreen Image="Assets\SplashScreen.png" />
-      </uap:VisualElements>
-    </Application>
-  </Applications>
-  <Capabilities>
-    <Capability Name="internetClient" />
-    <Capability Name="privateNetworkClientServer" />
-    <DeviceCapability Name="location" />
-    <DeviceCapability Name="proximity" />
-  </Capabilities>
-</Package>
