+++ src/GeocodeAndRoutingOnMouseMove/LocalNetworkSample.WinUI/Assets/Logo.scale-100.png
