+++ src/GeocodeAndRoutingOnMouseMove/LocalNetworkSample.WinUI/Assets/SmallLogo.scale-100.png
