+++ src/GeocodeAndRoutingOnMouseMove/LocalNetworkSample.WinUI/Assets/SplashScreen.scale-100.png
