+++ src/GeocodeAndRoutingOnMouseMove/LocalNetworkSample.WinUI/Assets/StoreLogo.scale-100.png
