+++ src/GeocodeAndRoutingOnMouseMove/LocalNetworkSample.WinUI/Controls/RouteDirectionView.xaml
