+++ src/GeocodeAndRoutingOnMouseMove/LocalNetworkSample.WinUI/Controls/RouteDirectionView.xaml.cs
@@ -7,13 +7,13 @@ using System.Linq;
 using System.Runtime.InteropServices.WindowsRuntime;
 using Windows.Foundation;
 using Windows.Foundation.Collections;
-using Windows.UI.Xaml;
-using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Controls.Primitives;
-using Windows.UI.Xaml.Data;
-using Windows.UI.Xaml.Input;
-using Windows.UI.Xaml.Media;
-using Windows.UI.Xaml.Navigation;
+using Microsoft.UI.Xaml;
+using Microsoft.UI.Xaml.Controls;
+using Microsoft.UI.Xaml.Controls.Primitives;
+using Microsoft.UI.Xaml.Data;
+using Microsoft.UI.Xaml.Input;
+using Microsoft.UI.Xaml.Media;
+using Microsoft.UI.Xaml.Navigation;
 
 // The User Control item template is documented at http://go.microsoft.com/fwlink/?LinkId=234236
 
@@ -47,10 +47,10 @@ namespace LocalNetworkSample.Controls
 		{
 			if (direction == null)
 			{
-				LayoutRoot.Visibility = Windows.UI.Xaml.Visibility.Collapsed;
+				LayoutRoot.Visibility =Microsoft.UI.Xaml.Visibility.Collapsed;
 				return;
 			}
-			LayoutRoot.Visibility = Windows.UI.Xaml.Visibility.Visible;
+			LayoutRoot.Visibility =Microsoft.UI.Xaml.Visibility.Visible;
 			LayoutRoot.DataContext = direction;
 			var d = LinearUnits.Miles.FromMeters(direction.Length);
 			if (d == 0)
