--- src/GeocodeAndRoutingOnMouseMove/packages/repositories.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<repositories>
-  <repository path="..\LocalNetworkSample.Desktop\packages.config" />
-  <repository path="..\LocalNetworkSample.WinStore\packages.config" />
-</repositories>
